requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11 